feat: form submit logic and protection
